Inline text input modal types
