--- apps/frontend/app/components/SettingsListInput/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SettingsListInput';
-export type { SettingsListInputProps } from './types';
